Add work_with_custom_tests
