+++ developers/academy/py/starter_custom_vectors/101_setup_weaviate/20_create_instance/10_create_wcs.mdx
@@ -5,7 +5,7 @@ title: "Option 1: A cloud WCS instance"
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import FilteredTextBlock from '@site/src/components/Documentation/FilteredTextBlock';
-import PyCode from '!!raw-loader!../_snippets/connect.py';
+import PyCode from '!!raw-loader!../../_snippets/101_connect.py';
 
 Here, you will create a Weaviate Cloud Service (WCS) instance. WCS is a fully managed Weaviate instance that runs in the cloud. It's a great way to get started with Weaviate, as it requires no installation or maintenance.
