+++ developers/academy/py/starter_custom_vectors/101_setup_weaviate/20_create_instance/20_create_docker.mdx
@@ -5,7 +5,7 @@ title: "Option 2: A local Docker instance"
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import FilteredTextBlock from '@site/src/components/Documentation/FilteredTextBlock';
-import PyCode from '!!raw-loader!../_snippets/connect.py';
+import PyCode from '!!raw-loader!../../_snippets/101_connect.py';
 
 :::note Have you already created a Weaviate instance?
 If you have [created a cloud instance](./10_create_wcs.mdx) of Weaviate, you can skip this page and continue with [Communicate with Weaviate](../30_communicate.mdx).
