+++ developers/academy/py/starter_custom_vectors/101_setup_weaviate/30_communicate.mdx
@@ -5,7 +5,7 @@ title: Communicate with Weaviate
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import FilteredTextBlock from '@site/src/components/Documentation/FilteredTextBlock';
-import PyCode from '!!raw-loader!./_snippets/connect.py';
+import PyCode from '!!raw-loader!../_snippets/101_connect.py';
 
 Here, we'll perform basic operations to communicate with Weaviate using the Python client library.
