+++ developers/academy/py/starter_custom_vectors/102_object_collections/20_create_collection.mdx
@@ -5,7 +5,7 @@ title: Create a collection
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import FilteredTextBlock from '@site/src/components/Documentation/FilteredTextBlock';
-import PyCode from '!!raw-loader!./_snippets/20_create_collection.py';
+import PyCode from '!!raw-loader!../_snippets/102_collections.py';
 
 Weaviate stores data in "collections". A collection is a set of objects that share the same data structure. In our movie database, we might have a collection of movies, a collection of actors, and a collection of reviews.
