+++ developers/academy/py/starter_custom_vectors/102_object_collections/25_generate_vectors.mdx
@@ -5,7 +5,7 @@ title: (Bonus) Generate vectors
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import FilteredTextBlock from '@site/src/components/Documentation/FilteredTextBlock';
-import PyCode from '!!raw-loader!./_snippets/25_generate_embeddings.py';
+import PyCode from '!!raw-loader!../_snippets/102_collections.py';
 
 Since we are using custom vectors, we need to generate them ourselves.
