+++ developers/academy/py/starter_custom_vectors/102_object_collections/30_import_data.mdx
@@ -5,7 +5,7 @@ title: Import data
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import FilteredTextBlock from '@site/src/components/Documentation/FilteredTextBlock';
-import PyCode from '!!raw-loader!./_snippets/30_import_data.py';
+import PyCode from '!!raw-loader!../_snippets/102_collections.py';
 
 ## <i class="fa-solid fa-code"></i> Code
