+++ developers/academy/py/starter_custom_vectors/103_object_searches/10_vector.mdx
@@ -5,7 +5,7 @@ title: Vector search
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import FilteredTextBlock from '@site/src/components/Documentation/FilteredTextBlock';
-import PyCode from '!!raw-loader!./_snippets/10_vector.py';
+import PyCode from '!!raw-loader!../_snippets/103_10_vector.py';
 
 In this scenario, we've added data objects and our own vectors. Accordingly, any similarity searches will also require a vector input. This can be done with a `near vector` query.
