+++ developers/academy/py/starter_custom_vectors/103_object_searches/20_keyword_hybrid.mdx
@@ -5,7 +5,7 @@ title: Keyword & Hybrid search
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import FilteredTextBlock from '@site/src/components/Documentation/FilteredTextBlock';
-import PyCode from '!!raw-loader!./_snippets/searches.py';
+import PyCode from '!!raw-loader!../_snippets/103_20_searches.py';
 
 You can also perform keyword (BM25) searches to find items based on their keyword similarity, or hybrid searches that combine BM25 and semantic/vector searches.
