+++ developers/academy/py/starter_custom_vectors/103_object_searches/30_filters.mdx
@@ -5,7 +5,7 @@ title: Filters
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import FilteredTextBlock from '@site/src/components/Documentation/FilteredTextBlock';
-import PyCode from '!!raw-loader!./_snippets/searches.py';
+import PyCode from '!!raw-loader!../_snippets/103_20_searches.py';
 
 Filters can be used to precisely refine search results. You can filter by properties as well as metadata, and you can combine multiple filters with `and` or `or` conditions to further narrow down the results.
