+++ developers/academy/py/starter_custom_vectors/104_object_rag/10_setup.mdx
@@ -5,7 +5,7 @@ title: "RAG: Overview"
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import FilteredTextBlock from '@site/src/components/Documentation/FilteredTextBlock';
-import PyCode from '!!raw-loader!../102_object_collections/_snippets/20_create_collection.py';
+import PyCode from '!!raw-loader!../_snippets/102_collections.py';
 
 ### <i class="fa-solid fa-chalkboard"></i> Motivation
