+++ developers/academy/py/starter_custom_vectors/104_object_rag/20_single_prompt.mdx
@@ -5,7 +5,7 @@ title: "'Single prompt' generation"
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import FilteredTextBlock from '@site/src/components/Documentation/FilteredTextBlock';
-import PyCode from '!!raw-loader!./_snippets/rag.py';
+import PyCode from '!!raw-loader!../_snippets/104_rag.py';
 
 A 'single prompt' generation wil perform RAG queries on each retrieved object. This is useful when you want to transform each object separately, with the same prompt.
