+++ developers/academy/py/starter_custom_vectors/104_object_rag/30_grouped_task.mdx
@@ -5,7 +5,7 @@ title: "'Grouped task' generation"
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import FilteredTextBlock from '@site/src/components/Documentation/FilteredTextBlock';
-import PyCode from '!!raw-loader!./_snippets/rag.py';
+import PyCode from '!!raw-loader!../_snippets/104_rag.py';
 
 A 'grouped task' generation wil perform RAG queries on the set of retrieved objects. This is useful when you want to transform the set of objects as a whole, with one prompt.
