+++ developers/academy/py/starter_custom_vectors/_snippets/101_connect.py
@@ -48,7 +48,6 @@ headers = {
 client = weaviate.connect_to_local(headers=headers)
 # END DockerAPIKeyInstantiation
 
-client.close()
 
 # PollLiveness
 assert client.is_live()  # This will raise an exception if the client is not live
